back/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.1.0
importlib-metadata==4.11.2
Markdown==3.3.6
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.1
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
xlrd==2.0.1
XlsxWriter==3.0.2
zipp==3.7.0
